% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-tag.R
\name{as_shiny_meta_tag}
\alias{as_shiny_meta_tag}
\title{Convert the CSP to a meta tag}
\usage{
as_shiny_meta_tag(csp)
}
\arguments{
\item{csp}{a content security policy.

This requires the shiny package.}
}
\description{
Convert the CSP to a meta tag
}
