% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\docType{data}
\name{content_security_policy}
\alias{content_security_policy}
\alias{default_src}
\alias{img_src}
\alias{media_src}
\alias{style_src}
\alias{script_src}
\alias{none}
\alias{unsafe_inline}
\alias{self}
\title{Define a content security policy}
\format{An object of class \code{csp_option} of length 1.}
\usage{
content_security_policy(...)

default_src(...)

img_src(...)

media_src(...)

style_src(...)

script_src(...)

none

unsafe_inline

self
}
\arguments{
\item{...}{CSP directives}
}
\description{
Define a content security policy
}
\examples{
# this is a default CSP that works with shiny:
csp <- content_security_policy(
  default_src(self),
  script_src(unsafe_inline),
  style_src(unsafe_inline),
  img_src(self, "data:")
)

}
\keyword{datasets}
